requirement2.txt
asgiref==3.7.2
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Flask==3.0.2
Flask-Cors==4.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysqlclient==2.2.4
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
tzdata==2023.4
Werkzeug==3.0.1
